chore: move viz to plot
