+++ wandb/__init__.py
@@ -76,8 +76,7 @@ from wandb.data_types import JoinedTable
 
 from wandb.wandb_agent import agent
 
-from wandb.sdk.lib.viz import visualize
-from wandb import plot
+from wandb.plot.viz import visualize
 from wandb.integration.sagemaker import sagemaker_auth
 from wandb.sdk.internal import profiler
