+++ wandb/sdk/internal/tb_watcher.py
@@ -12,9 +12,9 @@ from typing import TYPE_CHECKING, Any, Dict, List, Optional
 
 import wandb
 from wandb import util
+from wandb.plot.viz import CustomChart
 from wandb.sdk.interface.interface import GlobStr
 from wandb.sdk.lib import filesystem
-from wandb.sdk.lib.viz import CustomChart
 
 from . import run as internal_run
